Правки, разделение на модули
